Use float32 arrays for SICOR's retrieval maps to avoid GDAL float16 incompatibility issues.

+++ enpt/processors/atmospheric_correction/atmospheric_correction.py
@@ -391,11 +391,11 @@ class AtmosphericCorrector(object):
         if land_additional_results:
             # TODO: Consider to also join SICOR's retrieval uncertainty measures
             enmap_ImageL1.swir.sicor_cwv = \
-                np.nan_to_num(land_additional_results['cwv_model'], nan=-9999).astype(np.float16)
+                np.nan_to_num(land_additional_results['cwv_model'], nan=-9999).astype(np.float32)
             enmap_ImageL1.swir.sicor_liq = \
-                np.nan_to_num(land_additional_results['liq_model'], nan=-9999).astype(np.float16)
+                np.nan_to_num(land_additional_results['liq_model'], nan=-9999).astype(np.float32)
             enmap_ImageL1.swir.sicor_ice = \
-                np.nan_to_num(land_additional_results['ice_model'], nan=-9999).astype(np.float16)
+                np.nan_to_num(land_additional_results['ice_model'], nan=-9999).astype(np.float32)
 
         # join additional results from ACwater/Polymer
         if water_additional_results and self.cfg.polymer_additional_results:
